requirements.txt
albucore==0.0.23
albumentations==2.0.5
annotated-types==0.7.0
awsebcli==3.22.0
blessed==1.20.0
boto3==1.35.99
botocore==1.35.99
cement==2.10.14
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.1
colorama==0.4.6
filelock==3.18.0
Flask==2.0.3
flask-cors==5.0.1
fsspec==2025.3.2
idna==3.10
itsdangerous==2.1.2
Jinja2==3.1.1
jmespath==1.0.1
joblib==1.4.2
MarkupSafe==2.1.1
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
opencv-python-headless==4.11.0.86
packaging==24.2
pathspec==0.10.1
pillow==11.2.1
pydantic==2.11.3
pydantic_core==2.33.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.10.4
scipy==1.15.2
semantic-version==2.10.0
setuptools==78.1.0
simsimd==6.2.1
six==1.17.0
stringzilla==3.12.4
sympy==1.13.1
termcolor==2.5.0
torch==2.6.0
torchvision==0.21.0
triton==3.2.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==1.26.20
wcwidth==0.2.13
Werkzeug==2.1.0
flask-cors
gunicorn

